refactor(gui): migrate frankenterm-gui to wgpu/mlua/bitflags upgrades
Part of the 2026-06-14/15 workspace dependency-upgrade batch. Adapts the
`crates/frankenterm-gui` crate to the upgraded rendering and scripting
dependency APIs bumped in the manifest commit.

Migration highlights:

  - wgpu / webgpu wrapper: surface-error classification now matches on
    the project's `webgpu::WebGpuSurfaceTextureError` wrapper rather than
    raw `wgpu::SurfaceError` variants; render-state, draw, and webgpu
    paths updated for the new texture/surface API
    (renderstate.rs, termwindow/render/draw.rs, termwindow/webgpu.rs,
    termwindow/mod.rs, scripting/mod.rs).
  - mlua: drop the explicit `'lua` lifetime parameter from
    `UserDataFields`/`UserDataMethods` impls in the GUI scripting bindings
    (termwindow/mod.rs, scripting/mod.rs, scripting/guiwin.rs).
  - bitflags: customglyph block-glyph flag API update.
  - clippy: platform dead-code gating
    (`#[cfg_attr(not(target_os = "macos"), allow(dead_code))]`) plus
    rustfmt reflow across commands, overlays (confirm/prompt/selector),
    termwindow background/box_model/mouseevent/render/pane, and main.

Compile-compatibility ports only; no rendering-behavior change intended.

+++ crates/frankenterm-gui/src/termwindow/background.rs
@@ -54,8 +54,8 @@ impl CachedGradient {
         }
 
         // Map t which is in range [a, b] to range [c, d]
-        fn remap(t: f64, a: f64, b: f64, c: f64, d: f64) -> f64 {
-            (t - a) * ((d - c) / (b - a)) + c
+        fn remap(t: f64, a: f64, b: f64, c: f32, d: f32) -> f32 {
+            ((t - a) * (f64::from(d - c) / (b - a)) + f64::from(c)) as f32
         }
 
         let (dmin, dmax) = grad.domain();
@@ -145,7 +145,7 @@ impl CachedGradient {
                     };
 
                     let t = (nx + (x - cx).powi(2) + (ny + y - cy).powi(2)).sqrt() / radius;
-                    *pixel = to_pixel(grad.at(t));
+                    *pixel = to_pixel(grad.at(t as f32));
                 }
             }
         }
